* develop: (24 commits)
  Disabled "review by" check as it gets in the way for local branches. Disabled msi addition/removal checks as they fail too often (e.g. for files meant for only certain platforms).
  Win: replaced auto completion images. Changed file names on the way. Win: updated msi creation script for changed images. Linux: updated CMakeList.txt too.
  Review of Win project settings. - Removed outdated settings, corrected inheritance. - Added wb.query.analysis files (python) to project.
  Added validation to prevent module initialization from being called more than once. Fix for Bug #19313268: DOUBLE MENU INFO IN SCHEMA RIGHT CLICK
  Small changes in preparation for better utf-8 support in file names for model files. - zip_fdopen for writing the model zip doesn't work well due to a problem in libzip zip_close(). So, the code is there but yet out commented. - Changed #ifdef checks for newer libzip datastructs as this doesn't work. Instead use a compiler switch that exists only on newer libzip versions.
  Fix build error on OL6, where Gtk::EntryBuffer doesn't exist yet
  Added some ignores for git
  Windows:TreeNodeView: Stores the clicked column on the backend. Performance Schema Reports: Fixes context menu reset to prevent submenus being shown on the top left corner of the aplication.
  - backend: wb_module: add info to wb_log if system is running in FIPS mode, also store full env vars when debug3 lvl is turned on  - backend: wb_tunel: don't start tunnel if connection is not ssh based  - close #20236376 MYSQL WORKBENCH FAILS WITH \"CANNOT START SSH TUNNEL MANAGER\" WITH FIPS MODE
  - fixed syntax error in wb_admin_export.py python file (patch sended by Christian Hesse <mail@eworm.de>)
  For newly created triggers their SQL was not written to the trigger grt object until the user edited it the first time. That caused trouble when deleting triggers, that where created by never edited. Now the code is written on creation. Additionally the sql code was made complete to not cause a syntax error already on init (which doesn't work well with our normal error handling). All the processing, error checking etc. kicks in when the user starts editing.
  close #19802579: TYPE NOT SHOW FOR SPATIAL INDEX, fix cast warning and truncation warning text
  Fixing solution and project settings for fabric interface - Adding Release_OSS configuration. - Properly configuring Release_OSS/x64 solution to build fabric interface in 64 bits. - Correcting preprocessor define to export fabric interface functions.
  - python: sshtunnel: change var names so they'll not hide argument vars  review by Mike [jabber]
  Fixing Windows MEB Helper Scheduling Logic - Fixing Weekly/Montly scheduling logic.
  - backend: binary_editor: close form after pressing apply  - linux: frontend: grid_view: check if pasted text wasn't truncated if so, display warning  - linux: frontend: recordset_view: add missing initialization for variable  - close #14383416 MYSQL WORKBENCH TRIMS LONGTEXT FIELDS  review by rennox [revid 7550]
  Fix for grant parsing logic that is letting MEB unusable. - Added rule so an EOF token is added at the end of the GRANT statement. - Fixed user definition parsing to ensure only non keyword ids are quoted.
  - backend: autocomplete: remove g_thread_exit, as it's useless because of return null, add additional info to g_thread_join in create_thread  rev by Mike [revid: 7496]
  - backend: autocomplete: use RecTryLock to find out if we can obtain shutdown_mutex instead of exception throw
  - backend: query_side_palette,wb_sql_editor_form, wb_sql_editor_tree_controller, recordset:  	add description for dispatcher task  backend: autocomplete: instead of lock immediatelly, try to lock so we will not end with deadlock caused by shutdown() call
  ...

+++ plugins/wb.admin/frontend/wb_admin_perfschema_reports.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2013, 2014, Oracle and/or its affiliates. All rights reserved.
+# Copyright (c) 2013, 2015, Oracle and/or its affiliates. All rights reserved.
 #
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License as
@@ -361,23 +361,26 @@ class PSHelperViewTab(mforms.Box):
 
     def _header_menu_will_show(self, parent):
         column = self._tree.get_clicked_header_column()
-        
-        self._hmenu.remove_all()
 
-        item = self._hmenu.add_item_with_title("Set Display Unit", lambda: None, "change_unit")
-        unit = self._column_units[column]
-        if unit in time_units:
-            for label in time_units:
-                i = item.add_item_with_title(label, lambda self=self, column=column, label=label: self._change_unit(column, label), label)
-                if unit == label:
-                    i.set_checked(True)
-        elif unit in byte_units:
-            for label in byte_units:
-                i = item.add_item_with_title(label, lambda self=self, column=column, label=label: self._change_unit(column, label), label)
-                if unit == label:
-                    i.set_checked(True)
-        else:
-            item.set_enabled(False)
+        # This reset should onle be done when the Context Menu will be initially shown.
+        # On submenus should be skipped or they will be shown on the top left corner of the UI
+        if parent is None:
+            self._hmenu.remove_all()
+
+            item = self._hmenu.add_item_with_title("Set Display Unit", lambda: None, "change_unit")
+            unit = self._column_units[column]
+            if unit in time_units:
+                for label in time_units:
+                    i = item.add_item_with_title(label, lambda self=self, column=column, label=label: self._change_unit(column, label), label)
+                    if unit == label:
+                        i.set_checked(True)
+            elif unit in byte_units:
+                for label in byte_units:
+                    i = item.add_item_with_title(label, lambda self=self, column=column, label=label: self._change_unit(column, label), label)
+                    if unit == label:
+                        i.set_checked(True)
+            else:
+                item.set_enabled(False)
 
 
     def _tree_column_resized(self, column):
